requirements.txt
asgiref==3.5.2
black==23.3.0
click==8.1.3
Django==4.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.1.1
flake8==6.0.0
flake8-commas==2.1.0
flake8-import-order==0.18.2
flake8-variables-names==0.0.5
iniconfig==2.0.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.2.0
pluggy==1.0.0
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pyflakes==3.0.1
PyJWT==2.6.0
pytest==7.3.1
pytest-django==4.5.2
pytz==2022.2.1
sqlparse==0.4.2
tomli==2.0.1
typing-extensions==4.5.0
tzdata==2022.2
